Fix: New database schema approach, divided by domains.
