+++ src/database/drizzle/0014_conscious_mesmero.sql
@@ -1,0 +1,12 @@
+ALTER TABLE "account" DISABLE ROW LEVEL SECURITY;--> statement-breakpoint
+ALTER TABLE "session" DISABLE ROW LEVEL SECURITY;--> statement-breakpoint
+ALTER TABLE "user" DISABLE ROW LEVEL SECURITY;--> statement-breakpoint
+ALTER TABLE "verification" DISABLE ROW LEVEL SECURITY;--> statement-breakpoint
+DROP TABLE "account" CASCADE;--> statement-breakpoint
+DROP TABLE "session" CASCADE;--> statement-breakpoint
+DROP TABLE "user" CASCADE;--> statement-breakpoint
+DROP TABLE "verification" CASCADE;--> statement-breakpoint
+ALTER TABLE "tasks_properties_values" RENAME COLUMN "value_timestamp" TO "valueTimestamp";--> statement-breakpoint
+ALTER TABLE "projects" ADD COLUMN "created_by_user" text NOT NULL;--> statement-breakpoint
+ALTER TABLE "projects" ADD CONSTRAINT "projects_created_by_user_user_id_fk" FOREIGN KEY ("created_by_user") REFERENCES "public"."user"("id") ON DELETE no action ON UPDATE no action;--> statement-breakpoint
+DROP TYPE "public"."role";
