+++ src/database/drizzle/meta/_journal.json
@@ -99,6 +99,13 @@
       "when": 1773268585256,
       "tag": "0013_living_trauma",
       "breakpoints": true
+    },
+    {
+      "idx": 14,
+      "version": "7",
+      "when": 1773328537018,
+      "tag": "0014_conscious_mesmero",
+      "breakpoints": true
     }
   ]
 }
