+++ src/database/schema/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './auth'
+export * from './relations'
