+++ src/database/schema/index.ts
@@ -1,4 +1,4 @@
-export * from '../helpers';
-export * from './project';
-export * from './tasks';
-export * from './auth'
+export * from './projects'
+export * from './stages'
+export * from './tags'
+export * from './tasks'
