+++ src/database/schema/projects/index.ts
@@ -1,0 +1,4 @@
+export * from "./projects";
+export * from "./categoriesOfProjects";
+export * from "./projectTags";
+export * from "./relations";
