+++ src/database/schema/stages/index.ts
@@ -1,0 +1,2 @@
+export * from './stages'
+export * from './relations'
