+++ src/database/schema/tags/index.ts
@@ -1,0 +1,3 @@
+export * from "./tags";
+export * from "./tagTypes";
+export * from "./relations";
