+++ src/database/schema/tasks/index.ts
@@ -1,0 +1,3 @@
+export * from './tasks'
+export * from './tasksProperties'
+export * from './relations'
